Add xrotation to EllipseArc definition
It's probably a redundant member in strict sense. However the 2D
rendering APIs and 2D ellipse drawing workflows seem to assume a way to
rotate the ellipse without applying any external transform. In future if
we use this definition for a more stricter geometry kernel, we can just
ignore this member.